Correção OneShot
Correção do professor sobre a tarefa oneshot.
Quando a interrupção for iniciada, cria a task oneShot e carrega um novo contexto. Para executar uma nova tarefa naquele momento.

+++ sync.h
@@ -7,7 +7,7 @@
 // Padrão POSIX
 
 typedef struct sem {
-    int contador;
+    int8_t contador;
     uint8_t fila[MAX_USER_TASKS];
     uint8_t pos_input;
     uint8_t pos_output;
@@ -23,7 +23,7 @@ typedef struct mutex {
 } mutex_t;
 
 
-void sem_init(sem_t *sem, uint8_t valor);
+void sem_init(sem_t *sem, int8_t valor);
 void sem_wait(sem_t *sem);
 void sem_post(sem_t *sem);
